Alterando BigDecimal para Double e deletando duplicacao na pasta resources

+++ src/main/java/com/marmitech/Marmitech/DTO/RequestDTO/PedidoItemRequestDTO.java
@@ -2,6 +2,6 @@ package com.marmitech.Marmitech.DTO.RequestDTO;
 
 import java.math.BigDecimal;
 
-public record PedidoItemRequestDTO(int id, int pedidoId, int produtoId, int quantidade, BigDecimal precoUnitarioPedido) {
+public record PedidoItemRequestDTO(int id, int pedidoId, int produtoId, int quantidade, Double precoUnitarioPedido) {
 
 }
