+++ src/main/java/com/marmitech/Marmitech/DTO/ResponseDTO/PedidoItemResponseDTO.java
@@ -1,7 +1,5 @@
 package com.marmitech.Marmitech.DTO.ResponseDTO;
 
-import java.math.BigDecimal;
-
-public record PedidoItemResponseDTO(int id, int pedidoId, int produtoId, String clienteNome,String produtoNome, int quantidade, BigDecimal precoUnitarioPedido, BigDecimal subtotal) {
+public record PedidoItemResponseDTO(int id, int pedidoId, int produtoId, String clienteNome, String produtoNome, int quantidade, Double precoUnitarioPedido, Double subtotal) {
 
 }
